Use gpages_build.sh from Collaborne/collaborne-build-tools
The idea here is to use bower as a really stupid git client, and just
provide us with gpages_build.sh (and other stuff) on travis. We don't really
need gpages_build.sh in this repository, and maintaining the countless copies
gets really ugly.

collaborne-build-tools is a private repository, but this is *fine* for OSS, as
you don't need this dependency when working on the component or consuming it.
We only need travis to be able to get it. As bower tends to be nicely verbose
we do need to silence it though when accessing the repository.