Start with a better default for the locale
Instead of defaulting to 'en', we now use the navigator.language value.
As this might be an arbitrarily esoteric language we do now provide a property
for a 'fallback locale', which defaults to 'en'.

Effectively this should be a no-op for now: the navigator language tends to be a
value such as 'en-US', but our component doesn't understand that, and so will always go to the
fallback.