Fix the tests after 88e80f871bb657f6f1fe25d7f35c8eea4c3635e0
We now calculate the default locale again per-element, but prefer
the global locale if there is at least one other element with this
behavior and its locale had been changed.

When the last element is detached we reset the global locale again.

While there: fix the listener logic, we don't need to check whether the locale
differs (that is what polymer does anyways), but we should avoid setting it back
on our own element.

I believe there is still an issue, as we would now mass-cascade the locale setting,
each element trying to tell every other element about it.